created navbar and enabled react routing
