+++ Frontend/monch/src/index.js
@@ -1,12 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import { BrowserRouter } from "react-router-dom";
+import { BrowserRouter , Routes, Route, } from "react-router-dom";
+import Users from "./routes/Users"
+import Reviews from './routes/Reviews';
+import Resturants from './routes/Resturants';
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <BrowserRouter>
-  <App />
+  <Routes>
+  <Route path="/" element={<App />} />
+  <Route path="/users" element={<Users />} />
+  <Route path="/reviews" element={<Reviews />} />
+  <Route path="/resturants" element={<Resturants />} />
+  </Routes>
   </BrowserRouter>
 );
