+++ Frontend/monch/src/routes/Resturants.js
@@ -1,0 +1,13 @@
+import Navbar from "../components/NavBar";
+
+function Resturant() {
+    return (
+        <div>
+        <h1>Resturant</h1>
+        <Navbar></Navbar>
+        </div>
+
+    );
+}
+
+export default Resturant
