+++ Frontend/monch/src/routes/Reviews.js
@@ -1,0 +1,13 @@
+import Navbar from "../components/NavBar";
+
+function Reviews() {
+    return (
+        <div>
+        <h1>Reviews</h1>
+        <Navbar></Navbar>
+        </div>
+
+    );
+}
+
+export default Reviews
